Implement InitModalStyle()
I will be using a couple more modal windows after this one. So, it would
be nice to have one central way to initialise the style for those
windows.

This would also allow me to change the style of the modal windows when
the theme changes without having to rely on modifying the Draw() method.