Change: ColorScheme type: struct -> map...
...and the dark and light colorschemes are now available through another
map (ColorSchemes) as keys "dark" and "light".

I did this because the previous `SetTheme()` was incomplete in it's
implementation: it would reset the accent of the application back to
purple. Since the ColorScheme was stored as a struct and Accent as
`tcell.Color`, we could not get the same accent from the new ColorScheme
without resorting to reflection, which I didn't want to do.